Implemented not-found.html and updated GlobalExceptionHandle

+++ src/main/java/com/softserve/itacademy/service/impl/UserServiceImpl.java
@@ -5,6 +5,7 @@ import com.softserve.itacademy.repository.UserRepository;
 import com.softserve.itacademy.service.UserService;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -20,25 +21,32 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User create(User user) {
-            return userRepository.save(user);
+        return userRepository.save(user);
     }
 
     @Override
     public User readById(long id) {
         Optional<User> optional = userRepository.findById(id);
+        if (optional.isPresent()) {
             return optional.get();
+        }
+        throw new EntityNotFoundException("User with id " + id + " not found");
     }
 
     @Override
     public User update(User user) {
-            User oldUser = readById(user.getId());
-                return userRepository.save(user);
+        User oldUser = readById(user.getId());
+        return userRepository.save(user);
     }
 
     @Override
     public void delete(long id) {
         User user = readById(id);
+        if (user != null) {
             userRepository.delete(user);
+        } else {
+            throw new EntityNotFoundException("User with id " + id + " not found");
+        }
     }
 
     @Override
